requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
attrs==22.2.0
cffi==1.15.1
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2
inflection==0.5.1
jsonschema==4.17.3
Pillow==9.4.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
pytz==2023.2
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
